Changed Throttle into interface. Existing Throttle implementation renamed DualThrottle. Added NullThrottle implementation.
